split test file into multiple files
